test: extract runtime memory white-box suites
